+++ src/main/java/subway/controller/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package subway.step;
+package subway.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
