+++ src/main/java/subway/controller/StationController.java
@@ -1,4 +1,4 @@
-package subway.step;
+package subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
